requirements.txt
chromadb
Flask
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-vectorstores
langchain-groq